fix: replace volume-based rolloff factor with radius-based selector targeting
The regular value for volume in a /playsound command is 1.0 = 16 blocks. It's possible to increase it to increase the audible range (e.g. 2.0 = 32 blocks and so on), but decreasing it does *not* actually decrease the audible range, as you'd expect (e.g. 0.5 = 8 blocks). Instead, the sound is still audible within a 16-block range, but is softer overall.

So, the only to achieve a gradual rolloff less than 16 blocks, is by entirely limiting who will be able to hear the songs at all via target selection. As such, we can use the `distance` condition to play notes only to players in a certain range. This commit works with a base range of 9 adding ±3 blocks for lower and higher notes, giving an effective range between 6-12 blocks. This rolloff can be easily customized by tweaking the parameters of the sigmoid function used in the calculation.